\name{econdatar-package}
\alias{econdatar-package}
\alias{econdatar}
\docType{package}
\title{
\packageTitle{econdatar}
}
\description{
\packageDescription{econdatar}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{econdatar}
\packageIndices{econdatar}
This package provides and interface to the EconData database (http://www..econdata.co.za).

In short, EconData provides a single source for South African Macroeconomic Data. There exists a wealth of economic data spread over the public domain. This service aims to bring these disparate sources of data together into a single centralised repository. In order to make the data useful an open API and this package for R are also provided, allowing for automatable workflows.

The internals of the EconData database are based on the SDMX standard, therefore using this package to its full potential requires some very rudimentary knowledge of SDMX. We have provided a blog with in-depth tutorials that cover all the essentials of making the most of the data we provide, which can be found at http://www.randomsample.co.za.
}
\author{
%\packageAuthor{econdatar}

Maintainer: \packageMaintainer{econdatar}
}
%\references{
%~~ Literature or other references for background information ~~
%}
%%~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
%%~~ the R documentation directory ~~
\keyword{ package }
\seealso{
http://www.econdata.co.za
http://www.randomsample.co.za
}
%\examples{
%~~ simple examples of the most important functions ~~
%}
